refactor: change all images to webp
